* Properly split the message and targets so the unchanged case works. * In arg_condense, wipe redundant paths in recursive mode. * In _delta_dir, respect $targets->{''} that indicates this directory   is part of the targets.
git-svn-id: file:///home/chmrr/work/bps/complete-history/svk/orig-svn/trunk@798 4ccb8a5b-8ae0-0310-9b02-86531366991d

+++ lib/SVK/Command/Commit.pm
@@ -251,6 +251,7 @@ sub get_committable {
 	($self->{message}, $targets) =
 	    get_buffer_from_editor (loc('log message'), $self->target_prompt,
 				    undef, $file, $target->{copath}, $target->{targets});
+	die loc("No targets to commit.\n") if $#{$targets} < 0;
 	unlink $file;
     }
